moved components related to account login to the folder auth, session context  in context
